chore: Move AdsCreateVariableHandlesRequest to separate file
